feat: move matches building process from matches-build.ts to vite `buildStart`
